+++ README.md
@@ -13,8 +13,8 @@
 # 기능 목록
 - [O] 웰컴 타이틀 구현
 - [O] 유저 이름 TextField 구현
-- [ ] 이메일 TextField 구현
+- [O] 이메일 TextField 구현
 - [ ] 비밀번호 TextField 구현
 - [ ] 비밀번호 확인 TextField 구현
-- [ ] TextField 함수 그룹화
+- [O] TextField 함수 그룹화
 - [ ] 회원가입 버튼 구현
