+++ app/src/androidTest/java/nextstep/signup/study/SignUpComposeTest.kt
@@ -2,15 +2,9 @@ package nextstep.signup.study
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.text.TextStyle
@@ -18,6 +12,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
+import nextstep.signup.ui.component.TextFieldCompose
 import org.junit.Rule
 import org.junit.Test
 
@@ -51,6 +46,19 @@ class SignUpComposeTest {
             .onNodeWithText(placeholder)
             .assertExists()
     }
+
+    @Test
+    fun emailTextFieldTest() {
+        val placeholder = "Email"
+
+        composeTestRule.setContent {
+            TextFieldCompose(placeholder = placeholder)
+        }
+
+        composeTestRule
+            .onNodeWithText(placeholder)
+            .assertExists()
+    }
 }
 
 @Composable
@@ -68,20 +76,6 @@ fun TitleCompose(text: String) {
     )
 }
 
-@Composable
-fun TextFieldCompose(placeholder: String) {
-    var text by remember { mutableStateOf("") }
-
-    TextField(
-        modifier = Modifier
-            .fillMaxWidth().
-        testTag("textField"),
-        placeholder = { Text(placeholder) },
-        value = text,
-        onValueChange = { text = it }
-    )
-}
-
 @Preview(
     showBackground = true,
     backgroundColor = 0xFFFFFFFF
@@ -96,6 +90,12 @@ private fun TitleComposePreview() {
     backgroundColor = 0xFFFFFFFF
 )
 @Composable
-fun TextFieldComposePreview() {
+fun UserNameTextFieldComposePreview() {
     TextFieldCompose("UserName")
+}
+
+@Preview
+@Composable
+fun EmailTextFieldComposePreview() {
+    TextFieldCompose("Email")
 }
