+++ app/src/main/java/nextstep/signup/MainActivity.kt
@@ -3,26 +3,22 @@ package nextstep.signup
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import nextstep.signup.ui.component.TextFieldCompose
 import nextstep.signup.ui.theme.SignupTheme
 
 class MainActivity : ComponentActivity() {
@@ -45,8 +41,12 @@ fun Greeting() {
     ) {
         TitleCompose("Welcome to Compose \uD83D\uDE80")
 
-        Column(modifier = Modifier.padding(top = 62.dp)) {
+        Column(
+            modifier = Modifier.padding(top = 62.dp),
+            verticalArrangement = Arrangement.spacedBy(32.dp)
+        ) {
             TextFieldCompose("UserName")
+            TextFieldCompose("Email")
         }
     }
 }
@@ -62,20 +62,6 @@ fun TitleCompose(text: String) {
         style = TextStyle(
             fontWeight = FontWeight.Bold
         )
-    )
-}
-
-@Composable
-fun TextFieldCompose(placeholder: String) {
-    var text by remember { mutableStateOf("") }
-
-    TextField(
-        modifier = Modifier
-            .fillMaxWidth()
-            .testTag("user textField"),
-        placeholder = { Text(placeholder) },
-        value = text,
-        onValueChange = { text = it }
     )
 }
