+++ app/src/main/java/nextstep/signup/ui/component/NSTextField.kt
@@ -1,0 +1,26 @@
+package nextstep.signup.ui.component
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
+
+@Composable
+fun TextFieldCompose(placeholder: String) {
+    var text by remember { mutableStateOf("") }
+
+    TextField(
+        modifier = Modifier
+            .fillMaxWidth()
+            .testTag("user textField"),
+        placeholder = { Text(placeholder) },
+        value = text,
+        onValueChange = { text = it }
+    )
+}
